Limit dependencies of HA core (#2762)
