+++ homeassistant/helpers/config_validation.py
@@ -12,7 +12,7 @@ from homeassistant.const import (
     CONF_ALIAS, CONF_ENTITY_ID, CONF_VALUE_TEMPLATE, WEEKDAYS,
     CONF_CONDITION, CONF_BELOW, CONF_ABOVE, SUN_EVENT_SUNSET,
     SUN_EVENT_SUNRISE, CONF_UNIT_SYSTEM_IMPERIAL, CONF_UNIT_SYSTEM_METRIC)
-from homeassistant.helpers.entity import valid_entity_id
+from homeassistant.core import valid_entity_id
 import homeassistant.util.dt as dt_util
 from homeassistant.util import slugify
